Corrijo bug picoFino

+++ src/herramientas/desgastador/DesgastadorPorUsos.java
@@ -2,16 +2,16 @@ package herramientas.desgastador;
 
 public class DesgastadorPorUsos implements Desgastador {
 
-    private float durabilidadAQuitar;
+    private int durabilidadAQuitar;
     private int usos;
 
-    public DesgastadorPorUsos(int limiteUsos, float durabilidadHerramienta) {
+    public DesgastadorPorUsos(int limiteUsos, int durabilidadHerramienta) {
         usos = limiteUsos;
         durabilidadAQuitar = durabilidadHerramienta;
     }
 
     @Override
-    public float calcularDesgaste() {
+    public int calcularDesgaste() {
         if(usos == 0) return durabilidadAQuitar;
         usos -= 1;
         return 0;
